Initial support for using vcpkg
This seems to be the favored way for finding libraries under msvc, and
the `vcpkg` crate is a reverse dependency of various -sys crates.

It seems to have some issues as is. Vcpkg doesn't seem to have support
for finding a particular version of a package. And with this,
`glib-sys`, for instance, still fails to build without changes because it
needs to search for just `glib`, not `glib-2.0`.

Perhaps the solution is to add an optional `vcpkg_name` that can be
specified in toml? But in any case, this shows what's involved in
supporting vcpkg.