Honor manual requests for static linking
This allows a user to statically link a library even when not using
pkg-config as source. Previously, if SYSTEM_DEPS_<LIB>_NO_PKG_CONFIG was
set, the SYSTEM_DEPS_<LIB>_LINK=static configuration was ignored.

Since the user already went above and beyond to issue a strong intent of
static linking, no check is performed whether the file actually exists.
The linker will scream if something is wrong.